+++ components/common/Input/SearchInput/styles.css.ts
@@ -1,4 +1,4 @@
-import { style, styleVariants } from '@vanilla-extract/css'
+import { style } from '@vanilla-extract/css'
 
 export const container = style({
     position: 'relative',
@@ -8,7 +8,6 @@ export const container = style({
 })
 
 export const input = style({
-    width: '100%',
     paddingRight: '40px',
 })
 
@@ -20,16 +19,4 @@ export const iconWrapper = style({
     justifyContent: 'center',
     height: '48px',
     cursor: 'pointer',
-})
-
-export const sizeVariants = styleVariants({
-    small: {
-        width: '278px',
-    },
-    medium: {
-        width: '302px',
-    },
-    large: {
-        width: '358px',
-    },
 })
